feat(ec): switch to soroban-bn254-precompile and add set_soroban_bn254_backend
