+++ .github/workflows/ci.yml
@@ -10,7 +10,6 @@ env:
   CARGO_TERM_COLOR: always
 
 jobs:
-  # ---------- rustfmt ----------
   lint:
     runs-on: ubuntu-latest
     steps:
@@ -20,24 +19,19 @@ jobs:
           components: rustfmt
       - run: cargo fmt --all -- --check
 
-  # ---------- Tests (independent) ----------
   test:
     runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
 
-      # Rust toolchain
       - uses: dtolnay/rust-toolchain@stable
 
-      # --- Pin legacy Noir toolchain (Poseidon2 still in stdlib) -----------
       - name: Install Noir toolchain 1.0.0-beta.3
         uses: noir-lang/noirup@v0.1.2
         with:
           toolchain: "1.0.0-beta.3"
-      # --------------------------------------------------------------------
 
-      # Cache Cargo dependencies
       - uses: actions/cache@v3
         with:
           path: |
@@ -46,19 +40,17 @@ jobs:
             target
           key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
 
-      # ---------- Build all Noir circuits ----------
       - name: Generate proofs & VKs
         run: tests/build_circuits.sh
 
-      # ---------- Run Rust tests ----------
-      - name: Run tests
+      - name: Run tests (default features)
         run: cargo test --verbose
 
-      # ---------- Test alloc feature ----------
-      - name: Test alloc feature
-        run: cargo test alloc --verbose
+      - name: Run tests (std)
+        run: cargo test --features std --verbose
 
-      # ---------- Test std feature ----------
-      - name: Test std feature
-        run: cargo test std --verbose
+      - name: Run tests (soroban-bn254-precompile)
+        run: cargo test --features soroban-bn254-precompile --verbose
 
+      - name: Run tests (std + soroban-bn254-precompile)
+        run: cargo test --features "std soroban-bn254-precompile" --verbose
